Fix the bug when test case contains double quotes.

+++ src/tw/edu/ncku/csie/selab/jojs/Judger.java
@@ -81,7 +81,8 @@ public class Judger {
 
     public void cleanWorkingDirectory() {
         try {
-            FileUtils.forceDelete(workingDirectory);
+            if (workingDirectory.exists())
+                FileUtils.forceDelete(workingDirectory);
         } catch (IOException e) {
             e.printStackTrace();
         }
